rearrange the project into packages
