--- src/main/java/Event.java
@@ -1,26 +1,0 @@
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-class Event extends Task {
-    protected LocalDateTime from;
-    protected LocalDateTime to;
-
-    public Event(String description, String from, String to) {
-        super(description);
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
-        this.from = LocalDateTime.parse(from, formatter);
-        this.to = LocalDateTime.parse(to, formatter);
-    }
-
-    @Override
-    public String toString() {
-        DateTimeFormatter outputFormat = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mm a");
-        return "[E]" + super.toString() + " (from: " + from.format(outputFormat) + " to: " + to.format(outputFormat) + ")";
-    }
-
-    @Override
-    public String toFileFormat() {
-        DateTimeFormatter fileFormat = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
-        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from.format(fileFormat) + " | " + to.format(fileFormat);
-    }
-}
