--- src/main/java/Storage.java
@@ -1,53 +1,0 @@
-import java.io.*;
-import java.util.ArrayList;
-import java.util.List;
-
-public class Storage {
-
-    private final String filePath;
-
-    public Storage(String filePath) {
-        this.filePath = filePath;
-    }
-
-    public List<Task> load() throws IOException {
-        List<Task> tasks = new ArrayList<>();
-        File file = new File(filePath);
-        if (!file.exists()) {
-            return tasks; // No tasks to load
-        }
-
-        BufferedReader reader = new BufferedReader(new FileReader(file));
-        String line;
-        while ((line = reader.readLine()) != null) {
-            String[] parts = line.split(" \\| ");
-            Task task = switch (parts[0]) {
-                case "T" -> new Todo(parts[2]);
-                case "D" -> new Deadline(parts[2], parts[3]);
-                case "E" -> new Event(parts[2], parts[3], parts[4]);
-                default -> null;
-            };
-            if (task != null && parts[1].equals("1")) {
-                task.markAsDone();
-            }
-            tasks.add(task);
-        }
-        reader.close();
-        return tasks;
-    }
-
-    public void save(List<Task> tasks) throws IOException {
-        File directory = new File("./data");
-        if (!directory.exists()) {
-            directory.mkdir();
-        }
-
-        File file = new File(filePath);
-        BufferedWriter writer = new BufferedWriter(new FileWriter(file));
-        for (Task task : tasks) {
-            writer.write(task.toFileFormat());
-            writer.newLine();
-        }
-        writer.close();
-    }
-}
