--- src/main/java/TaskManagerException.java
@@ -1,5 +1,0 @@
-public class TaskManagerException extends Exception {
-    public TaskManagerException(String message) {
-        super(message);
-    }
-}
