--- src/main/java/Todo.java
@@ -1,15 +1,0 @@
-class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String toFileFormat() {
-        return "T | " + (isDone ? "1" : "0") + " | " + description;
-    }
-}
