--- src/main/java/Ui.java
@@ -1,60 +1,0 @@
-import java.util.Scanner;
-
-public class Ui {
-
-    private final Scanner scanner;
-
-    public Ui() {
-        scanner = new Scanner(System.in);
-    }
-
-    public void showWelcome() {
-        System.out.println("Hello! I'm ArsenBot");
-        System.out.println("What can I do for you?\n");
-    }
-
-    public void showLine() {
-        System.out.println("____________________________");
-    }
-
-    public void showError(String message) {
-        System.out.println(message);
-    }
-
-    public String readCommand() {
-        return scanner.nextLine();
-    }
-
-    public void showExit() {
-        System.out.println("Bye. Hope to see you again soon!\n");
-    }
-
-    public void showTaskAdded(Task task, int size) {
-        System.out.println("Got it. I've added this task:");
-        System.out.println("\t" + task);
-        System.out.println("Now you have " + size + " tasks in the list.");
-    }
-
-    public void showTaskDeleted(Task task, int size) {
-        System.out.println("Noted. I've removed this task:");
-        System.out.println(task);
-        System.out.println("Now you have " + size + " tasks in the list.");
-    }
-
-    public void showTaskMarked(Task task) {
-        System.out.println("Nice! I've marked this task as done:");
-        System.out.println(task);
-    }
-
-    public void showTaskUnmarked(Task task) {
-        System.out.println("OK, I've marked this task as not done yet:");
-        System.out.println(task);
-    }
-
-    public void showTaskList(TaskList tasks) {
-        System.out.println("Here are the tasks in your list:");
-        for (int i = 0; i < tasks.size(); i++) {
-            System.out.println((i + 1) + "." + tasks.get(i));
-        }
-    }
-}
