+++ src/main/java/arsenbot/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package arsenbot.command;
+
+import arsenbot.storage.Storage;
+import arsenbot.task.TaskList;
+import arsenbot.ui.Ui;
+
 public class ExitCommand extends Command {
 
     @Override
