+++ src/main/java/arsenbot/command/MarkCommand.java
@@ -1,11 +1,19 @@
+package arsenbot.command;
+
+import arsenbot.storage.Storage;
+import arsenbot.task.Task;
+import arsenbot.task.TaskList;
+import arsenbot.task.TaskManagerException;
+import arsenbot.ui.Ui;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private final int index;
 
-    public UnmarkCommand(String input) throws TaskManagerException {
+    public MarkCommand(String input) throws TaskManagerException {
         try {
-            this.index = Integer.parseInt(input.substring(7)) - 1;
+            this.index = Integer.parseInt(input.substring(5)) - 1;
         } catch (NumberFormatException e) {
             throw new TaskManagerException("Error: Invalid task number.");
         }
@@ -17,8 +25,8 @@ public class UnmarkCommand extends Command {
             throw new TaskManagerException("Error: Invalid task number.");
         }
         Task task = tasks.get(index);
-        task.markAsUndone();
-        ui.showTaskUnmarked(task);
+        task.markAsDone();
+        ui.showTaskMarked(task);
         storage.save(tasks.getTasks());
     }
 }
