+++ src/main/java/arsenbot/command/UnmarkCommand.java
@@ -1,9 +1,17 @@
+package arsenbot.command;
+
+import arsenbot.storage.Storage;
+import arsenbot.task.Task;
+import arsenbot.task.TaskList;
+import arsenbot.task.TaskManagerException;
+import arsenbot.ui.Ui;
+
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     private final int index;
 
-    public DeleteCommand(String input) throws TaskManagerException {
+    public UnmarkCommand(String input) throws TaskManagerException {
         try {
             this.index = Integer.parseInt(input.substring(7)) - 1;
         } catch (NumberFormatException e) {
@@ -17,8 +25,8 @@ public class DeleteCommand extends Command {
             throw new TaskManagerException("Error: Invalid task number.");
         }
         Task task = tasks.get(index);
-        tasks.deleteTask(index);
-        ui.showTaskDeleted(task, tasks.size());
+        task.markAsUndone();
+        ui.showTaskUnmarked(task);
         storage.save(tasks.getTasks());
     }
 }
