+++ src/main/java/arsenbot/task/TaskList.java
@@ -1,0 +1,37 @@
+package arsenbot.task;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TaskList {
+
+    private final List<Task> tasks;
+
+    public TaskList() {
+        tasks = new ArrayList<>();
+    }
+
+    public TaskList(List<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public void addTask(Task task) {
+        tasks.add(task);
+    }
+
+    public void deleteTask(int index) {
+        tasks.remove(index);
+    }
+
+    public Task get(int index) {
+        return tasks.get(index);
+    }
+
+    public int size() {
+        return tasks.size();
+    }
+
+    public List<Task> getTasks() {
+        return tasks;
+    }
+}
