+++ src/main/java/arsenbot/task/TaskManagerException.java
@@ -1,0 +1,7 @@
+package arsenbot.task;
+
+public class TaskManagerException extends Exception {
+    public TaskManagerException(String message) {
+        super(message);
+    }
+}
